% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_pnl.R
\name{aggregate_strategy}
\alias{aggregate_strategy}
\title{Calculate strategy pnl for single option}
\usage{
aggregate_strategy(option_measure, position, opening_price,
  multiplier = 100)
}
\arguments{
\item{position}{Double : position in that option}

\item{opening_price}{Double}

\item{multiplier}{Integer : option multiplier}

\item{pnl}{Double : pnl vector}
}
\description{
Calculate strategy pnl for single option
}
