% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_pnl.R
\name{compute_option_greeks}
\alias{compute_option_greeks}
\title{Compute selected option greek for given vol+price+time}
\usage{
compute_option_greeks(underlyer_space, ttm = self$time_to_expiration,
  implied_vol = self$implied_vol, greeks = c("delta", "gamma", "vega",
  "theta"))
}
\arguments{
\item{greek}{}
}
\description{
Compute selected option greek for given vol+price+time
}
