% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_pnl.R
\name{compute_strategy_pnl}
\alias{compute_strategy_pnl}
\title{Compute strategy pnl}
\usage{
compute_strategy_pnl(days_to_exp = 0, underlyer_pct_move = 0.1)
}
\arguments{
\item{days_to_exp}{Number of days to expiration, numeric}
}
\value{
List of pnl, breakevens, max profit and max loss
}
\description{
Compute strategy pnl
}
