% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_pnl.R
\name{create_strat}
\alias{create_strat}
\title{Create strategy from dataframe}
\usage{
create_strat(strat_dt, days_to_exp = 0)
}
\arguments{
\item{days_to_exp}{}
}
\value{

}
\description{
Create strategy from dataframe
}
