% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_pnl.R
\name{plot_strategy_pnl}
\alias{plot_strategy_pnl}
\title{Plot strategy pnl}
\usage{
plot_strategy_pnl(days_to_exp)
}
\arguments{
\item{days_to_exp}{Number of days until expiration for all legs in strategy}
}
\value{
plotly plot of strategy pnl
}
\description{
Plot strategy pnl
}
